Projet/requirements.txt
aws-cdk-lib==2.3.0
constructs>=10.0.0,<11.0.0
aws_cdk.aws_lambda_python_alpha==2.3.0
aws_cdk.aws_apigatewayv2_alpha==2.3.0
aws_cdk.aws_apigatewayv2_integrations_alpha==2.3.0
aws_cdk.aws_glue_alpha==2.3.0